requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mypy==1.7.1
ruff==0.1.6
langchain==1.2.4
langchain-anthropic==1.3.0
langchain-core==1.2.7
langgraph==1.0.6
pydantic==2.12.5
python-dotenv==1.2.1
